+++ ext_localconf.php
@@ -2,43 +2,43 @@
 defined('TYPO3_MODE') or die('Access denied.');
 
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addUserTSConfig('
-	options.saveDocNew.tx_seminars_seminars=1
+    options.saveDocNew.tx_seminars_seminars=1
 ');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addUserTSConfig('
-	options.saveDocNew.tx_seminars_speakers=1
+    options.saveDocNew.tx_seminars_speakers=1
 ');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addUserTSConfig('
-	options.saveDocNew.tx_seminars_attendances=1
+    options.saveDocNew.tx_seminars_attendances=1
 ');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addUserTSConfig('
-	options.saveDocNew.tx_seminars_sites=1
+    options.saveDocNew.tx_seminars_sites=1
 ');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addUserTSConfig('
-	options.saveDocNew.tx_seminars_organizers=1
+    options.saveDocNew.tx_seminars_organizers=1
 ');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addUserTSConfig('
-	options.saveDocNew.tx_seminars_payment_methods=1
+    options.saveDocNew.tx_seminars_payment_methods=1
 ');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addUserTSConfig('
-	options.saveDocNew.tx_seminars_event_types=1
+    options.saveDocNew.tx_seminars_event_types=1
 ');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addUserTSConfig('
-	options.saveDocNew.tx_seminars_checkboxes=1
+    options.saveDocNew.tx_seminars_checkboxes=1
 ');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addUserTSConfig('
-	options.saveDocNew.tx_seminars_lodgings=1
+    options.saveDocNew.tx_seminars_lodgings=1
 ');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addUserTSConfig('
-	options.saveDocNew.tx_seminars_foods=1
+    options.saveDocNew.tx_seminars_foods=1
 ');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addUserTSConfig('
-	options.saveDocNew.tx_seminars_target_groups=1
+    options.saveDocNew.tx_seminars_target_groups=1
 ');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addUserTSConfig('
-	options.saveDocNew.tx_seminars_categories=1
+    options.saveDocNew.tx_seminars_categories=1
 ');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addUserTSConfig('
-	options.saveDocNew.tx_seminars_skills=1
+    options.saveDocNew.tx_seminars_skills=1
 ');
 
 // Adds our custom function to a hook in \TYPO3\CMS\Core\DataHandling\DataHandler
@@ -135,3 +135,13 @@ $iconRegistry->registerIcon(
     $iconProviderClass,
     ['source' => 'EXT:seminars/Resources/Public/Icons/Registration.gif']
 );
+
+$iconRegistry->registerIcon(
+    'ext-seminars-wizard-icon',
+    $iconProviderClass,
+    ['source' => 'EXT:seminars/ext_icon.svg']
+);
+
+\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPageTSConfig('
+    <INCLUDE_TYPOSCRIPT: source="FILE:EXT:seminars/Configuration/TSconfig/ContentElementWizard.txt">
+');