+++ Classes/BackEnd/AbstractList.php
@@ -146,7 +146,7 @@ abstract class Tx_Seminars_BackEnd_AbstractList
             );
             $langDelete = $languageService->getLL('delete', 1);
             $result = '<a class="btn btn-default" href="' .
-                htmlspecialchars($this->page->doc->issueCommand($params)) .
+                htmlspecialchars(BackendUtility::getLinkToDataHandlerAction($params)) .
                 '" onclick="' . $confirmation . '">' .
                 '<img src="/' . ExtensionManagementUtility::siteRelPath('seminars') . 'Resources/Public/Icons/Garbage.gif' .
                 '" title="' . $langDelete . '" alt="' . $langDelete . '" class="deleteicon" />' .
@@ -316,7 +316,7 @@ abstract class Tx_Seminars_BackEnd_AbstractList
             }
 
             $result = '<a class="btn btn-default" href="' .
-                htmlspecialchars($this->page->doc->issueCommand($params)) . '">' .
+                htmlspecialchars(BackendUtility::getLinkToDataHandlerAction($params)) . '">' .
                 '<img src="/' . ExtensionManagementUtility::siteRelPath('seminars') . 'Resources/Public/Icons/' . $icon .
                 '" title="' . $langHide . '" alt="' . $langHide . '" class="hideicon" />' .
                 '</a>';