+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Drop support for PHP 5.5 (#64)
 
 ### Fixed
+- Drop usage of the deprecated extRelPath method (#76)
 - Update the content element wizard for TYPO3 8.7 (#72)
 - Make the PHPUnit test runner configurable (#71)